@@ -15,14 +15,14 @@ apply plugin: 'maven-publish'
 repositories {
 }
 
-version = '1.16.5-1.0.5'
+version = '1.17.1-1.0.6'
 group = 'magic-mirror'
 archivesBaseName = 'SBM-Magic-Mirror'
 
-sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '1.8'
+java.toolchain.languageVersion = JavaLanguageVersion.of(16)
 
 minecraft {
-    mappings channel: 'official', version: '1.16.5'
+    mappings channel: 'official', version: '1.17.1'
     runs {
         client {
             properties 'forge.logging.console.level': 'debug'
@@ -38,19 +38,21 @@ minecraft {
 }
 
 dependencies {
-    minecraft 'net.minecraftforge:forge:1.16.5-36.2.0'
+    minecraft 'net.minecraftforge:forge:1.17.1-37.0.12'
 }
 
+
 jar {
-    manifest {
-        attributes([
-                "Specification-Title"     : archivesBaseName,
-                "Specification-Vendor"    : "StrikerRocker",
-                "Specification-Version"   : version,
-                "Implementation-Title"    : archivesBaseName,
-                "Implementation-Version"  : version,
-                "Implementation-Vendor"   : "StrikerRocker",
-                "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")
-        ])
+    from('/') {
+        include 'LICENSE.txt'
     }
+    manifest.attributes([
+            "Specification-Title"     : archivesBaseName,
+            "Specification-Vendor"    : "StrikerRocker",
+            "Specification-Version"   : version,
+            "Implementation-Title"    : archivesBaseName,
+            "Implementation-Version"  : version,
+            "Implementation-Vendor"   : "StrikerRocker",
+            "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")
+    ])
 }
