@@ -1,27 +1,19 @@
 package io.github.strikerrocker.magicmirror.handler;
 
 import io.github.strikerrocker.magicmirror.MagicMirror;
-import io.github.strikerrocker.magicmirror.capability.IMirrorData;
-import io.github.strikerrocker.magicmirror.capability.MirrorData;
+import io.github.strikerrocker.magicmirror.capability.CapabilitySerializerProvider;
 import io.github.strikerrocker.magicmirror.config.ConfigCost;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.Direction;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.common.capabilities.Capability;
-import net.minecraftforge.common.capabilities.ICapabilityProvider;
-import net.minecraftforge.common.util.LazyOptional;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.entity.EntityTravelToDimensionEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.util.HashMap;
 import java.util.UUID;
-
 
 /**
  * Handles everything to do with the mirror activation and tracking
@@ -35,7 +27,6 @@ public class MirrorHandler {
 
     private static final ResourceLocation CAP_KEY = new ResourceLocation(MagicMirror.DOMAIN, "teleport_position");
 
-    private static final LazyOptional<IMirrorData> holder = LazyOptional.of(MirrorData::new);
 
     private MirrorHandler() {
     }
@@ -43,11 +34,11 @@ public class MirrorHandler {
     /**
      * Updates tracked information about the player
      */
-    public static void updateUserData(PlayerEntity player, ItemStack mirror) {
+    public static void updateUserData(Player player, ItemStack mirror) {
         get(player).update(player, mirror);
     }
 
-    public static EntityData get(PlayerEntity player) {
+    public static EntityData get(Player player) {
         final UUID id = player.getGameProfile().getId();
         if (!userData.containsKey(id)) {
             userData.put(player.getUUID(), new EntityData(player));
@@ -58,7 +49,7 @@ public class MirrorHandler {
     /**
      * Called to teleport the player to his set mirror location
      */
-    public static void teleport(PlayerEntity player) {
+    public static void teleport(Player player) {
         player.getCapability(MagicMirror.CAPABILITY_MIRROR).ifPresent(mirrorData -> {
             if (mirrorData.hasLocation()) {
                 mirrorData.getLocation().teleport(player);
@@ -72,24 +63,16 @@ public class MirrorHandler {
 
     @SubscribeEvent
     public static void onChangeDimension(EntityTravelToDimensionEvent event) {
-        if (event.getEntity() instanceof PlayerEntity) {
-            PlayerEntity player = (PlayerEntity) event.getEntity();
+        if (event.getEntity() instanceof Player player) {
             get(player).reset();
             player.getCapability(MagicMirror.CAPABILITY_MIRROR).ifPresent(iMirrorData -> iMirrorData.setLocation(null));
         }
     }
 
-
     @SubscribeEvent
     public static void attachCapabilities(AttachCapabilitiesEvent<Entity> event) {
-        if (event.getObject() instanceof PlayerEntity) {
-            event.addCapability(CAP_KEY, new ICapabilityProvider() {
-                @Nonnull
-                @Override
-                public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side) {
-                    return MagicMirror.CAPABILITY_MIRROR.orEmpty(cap, holder);
-                }
-            });
+        if (event.getObject() instanceof Player) {
+            event.addCapability(CAP_KEY, new CapabilitySerializerProvider());
         }
     }
 }
