@@ -1,10 +1,11 @@
 package io.github.strikerrocker.magicmirror.handler;
 
 import io.github.strikerrocker.magicmirror.MagicMirror;
-import net.minecraft.loot.LootEntry;
-import net.minecraft.loot.LootPool;
-import net.minecraft.loot.TableLootEntry;
-import net.minecraft.util.ResourceLocation;
+import io.github.strikerrocker.magicmirror.config.ConfigLoot;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.storage.loot.LootPool;
+import net.minecraft.world.level.storage.loot.entries.LootPoolEntryContainer;
+import net.minecraft.world.level.storage.loot.entries.LootTableReference;
 import net.minecraftforge.event.LootTableLoadEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -17,7 +18,7 @@ public class LootTableHandler {
     public static void registerLoot(LootTableLoadEvent evt) {
         String prefix = "minecraft:chests/";
         String name = evt.getName().toString();
-        if (name.startsWith(prefix)) {
+        if (name.startsWith(prefix) && ConfigLoot.enableAsDungeonLoot.get()) {
             String file = name.substring(name.indexOf(prefix) + prefix.length());
             switch (file) {
                 case "abandoned_mineshaft":
@@ -47,9 +48,9 @@ public class LootTableHandler {
                 .build();
     }
 
-    private static LootEntry.Builder<?> getInjectEntry(String name) {
+    private static LootPoolEntryContainer.Builder<?> getInjectEntry(String name) {
         ResourceLocation table = new ResourceLocation(MagicMirror.DOMAIN, "inject/" + name);
-        return TableLootEntry.lootTableReference(table)
+        return LootTableReference.lootTableReference(table)
                 .setWeight(1);
     }
 }
